add ImportTransactionsService

+++ src/services/DeleteTransactionService.ts
@@ -1,9 +1,17 @@
-// import AppError from '../errors/AppError';
+import { getCustomRepository } from 'typeorm';
+import TransactionsRepository from '../repositories/TransactionsRepository';
 
+// import AppError from '../errors/AppError';
+interface Request {
+  id: string;
+}
 class DeleteTransactionService {
-  public async execute(): Promise<void> {
+  public async execute({ id }: Request): Promise<void> {
     const transactionsRepository = getCustomRepository(TransactionsRepository);
-    await transactionsRepository.delete({ id });
+
+    const findTransaction = await transactionsRepository.findOne(id);
+
+    if (findTransaction) await transactionsRepository.delete({ id });
   }
 }
